A fix for Issue #4 (from fspeech):
Okay I see what is going on. Python on Windows writes to files with DOS style
line endings by default, which changes the byte counts of the index file. The
solution is to write the index file in binary mode so that no extra characters
are added to the file. Here is the one line (character) change to pycscope.py
and should make it work on Windows (it seems to be working for me).