Remove threads; fix strings-as-symbols
Remove support for threads as it was not useful and it was broken.

Apply fixes for pyflakes warnings.

Fix strings-as-symbols support as it was broken by a missing global
statement. We then removed the listing of the string as is and the
string marked off with double brackets to just using double brackets
to allow more characters across the screen in cscope.

Signed-off-by: Peter Portante <peter.portante@redhat.com>