Use PEP 263 encoding detection in Py3k.
For Py3K we can use tokenize.open to read source code files in the
preferred style (pep 263).

Trap any exception during source code parsing and continue on instead of
aborting the process.  For large file lists that include the python
standard library (and its tests) there are unavoidable exceptions of a
variety of sources and we don't want to break the entire DB generation
when this happens.